Separate synchronizer implementation from models
