+++ settings.gradle
@@ -49,3 +49,5 @@ include ':logger'
 include ':logger:aspect'
 include ':logger:annotations'
 include ':synchronizer'
+include ':synchronizer:model'
+include ':synchronizer:core'
