--- synchronizer/api/current.api
@@ -1,29 +1,0 @@
-// Signature format: 4.0
-package com.urlaunched.synchronizer {
-
-  public abstract class BaseSynchronizer {
-    ctor public BaseSynchronizer();
-    method public final suspend Object? emitCancelDelete(com.urlaunched.synchronizer.Synchronizable<?> value, kotlin.coroutines.Continuation<? super kotlin.Unit>);
-    method public final suspend Object? emitDelete(com.urlaunched.synchronizer.Synchronizable<?> value, kotlin.coroutines.Continuation<? super kotlin.Unit>);
-    method public final suspend Object? emitUpdate(com.urlaunched.synchronizer.Synchronizable<?> value, kotlin.coroutines.Continuation<? super kotlin.Unit>);
-    method public final kotlinx.coroutines.flow.MutableSharedFlow<com.urlaunched.synchronizer.Synchronizable<?>> getCancelDeleteModel();
-    method public final kotlinx.coroutines.flow.MutableSharedFlow<com.urlaunched.synchronizer.Synchronizable<?>> getDeletedModel();
-    method public final kotlinx.coroutines.flow.MutableSharedFlow<com.urlaunched.synchronizer.Synchronizable<?>> getUpdateModel();
-    method public inline <reified T extends com.urlaunched.synchronizer.Synchronizable<ID>, ID> java.util.List<? extends T> synchronize(java.util.List<? extends T>, kotlinx.coroutines.CoroutineScope viewModelScope);
-    method public inline <reified T extends com.urlaunched.synchronizer.Synchronizable<ID>, reified R extends com.urlaunched.synchronizer.Synchronizable<ID>, ID> java.util.List<? extends T> synchronize(java.util.List<? extends T>, kotlinx.coroutines.CoroutineScope viewModelScope, kotlin.reflect.KClass<R> relatedType, optional kotlin.jvm.functions.Function2<? super T,? super R,? extends T> map);
-    method public inline <reified T extends com.urlaunched.synchronizer.Synchronizable<ID>, ID> kotlinx.coroutines.flow.Flow<? extends androidx.paging.PagingData<T>> synchronize(kotlinx.coroutines.flow.Flow<? extends androidx.paging.PagingData<T>>, kotlinx.coroutines.CoroutineScope viewModelScope);
-    method public inline <reified T extends com.urlaunched.synchronizer.Synchronizable<ID>, reified R extends com.urlaunched.synchronizer.Synchronizable<ID>, ID> kotlinx.coroutines.flow.Flow<? extends androidx.paging.PagingData<T>> synchronize(kotlinx.coroutines.flow.Flow<? extends androidx.paging.PagingData<T>>, kotlinx.coroutines.CoroutineScope viewModelScope, kotlin.reflect.KClass<R> relatedType, optional kotlin.jvm.functions.Function2<? super T,? super R,? extends T> map);
-    method public suspend inline <reified T extends com.urlaunched.synchronizer.Synchronizable<ID>, ID> void synchronize(T, kotlin.jvm.functions.Function1<? super T,? extends kotlin.Unit> onUpdate);
-    method public suspend inline <reified T extends com.urlaunched.synchronizer.Synchronizable<ID>, reified R extends com.urlaunched.synchronizer.Synchronizable<ID>, ID> void synchronize(T, kotlin.jvm.functions.Function1<? super T,? extends kotlin.Unit> onUpdate, kotlin.reflect.KClass<R> relatedType, kotlin.jvm.functions.Function1<? super R,? extends T> map);
-    property public final kotlinx.coroutines.flow.MutableSharedFlow<com.urlaunched.synchronizer.Synchronizable<?>> cancelDeleteModel;
-    property public final kotlinx.coroutines.flow.MutableSharedFlow<com.urlaunched.synchronizer.Synchronizable<?>> deletedModel;
-    property public final kotlinx.coroutines.flow.MutableSharedFlow<com.urlaunched.synchronizer.Synchronizable<?>> updateModel;
-  }
-
-  public interface Synchronizable<T> {
-    method public T getId();
-    property public abstract T id;
-  }
-
-}
-
