+++ synchronizer/core/.gitignore
