+++ synchronizer/core/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=synchronizer-core
+POM_NAME=Data syncronizer core
+POM_PACKAGING=aar
+VERSION_NAME=LOCAL
