+++ synchronizer/core/proguard-rules.pro
