+++ synchronizer/core/src/main/AndroidManifest.xml
