--- synchronizer/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=synchronizer
-POM_NAME=Data syncronizer
-POM_PACKAGING=aar
-VERSION_NAME=LOCAL
