+++ synchronizer/model/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=synchronizer-model
+POM_NAME=Synchronizer models
+POM_PACKAGING=aar
+VERSION_NAME=LOCAL
