+++ synchronizer/model/src/main/java/com/urlaunched/android/synchonizer/model/Synchronizable.kt
@@ -1,0 +1,5 @@
+package com.urlaunched.android.synchonizer.model
+
+interface Synchronizable<T> {
+    val id: T
+}
