--- synchronizer/src/main/java/com/urlaunched/synchronizer/Synchronizable.kt
@@ -1,5 +1,0 @@
-package com.urlaunched.synchronizer
-
-interface Synchronizable<T> {
-    val id: T
-}
